refactor: move Spotify account to the Integration screen
Spotify is an account integration like Last.fm, ListenBrainz and Discord,
but it was filed under Backup & Restore, so connecting an account meant
opening the backup screen.

Moves the Spotify account entry, the cookie login sheet, the error dialog
and the WebView login plumbing to IntegrationScreen under its own
"External service" group. Settings search already pointed at the
Integration page for Spotify, so nothing else needs to change.
